add simple multi user

+++ controllers/api/deleteNoteHandler.js
@@ -2,9 +2,14 @@ import response from '../../helpers/response.js';
 import { Notes } from '../../model/model.js';
 
 const deleteNoteHandler = async (req, res) => {
+  const { name } = req.user;
   const { id } = req.params;
 
   try {
+    const foundNotes = await Notes.find({ username: name });
+
+    if (!foundNotes) throw Error();
+
     const result = await Notes.deleteOne({ _id: id });
 
     if (result.deletedCount) {
